hello analysis: starting to write some print helpers to get a reasonable volume of quick info as I frolic about.
I'm sort of disappointed in what the stdlib has for this.
(Either that or I'm just not finding the right incantations yet.)
The printer and formatter stuff is of course advanced, but...
has interesting restrictions, such as only working starting from
a certain list of top-level node types.  While I'm sure that's
very useful for the tools that those things are used to build,
it's not very useful to me right now in debugging and playing.