more notes.
some predate when I started playing with the ast packages,
and so may already look sillyly speculative.

importantly: separating which parts of the job there are,
and realizing we can (and should) focus on each separately.
Somewhat surprisingly, one is probably going to require a lot more
of a fine-grained opt-in approach than the other (and is only
even going to be conditionally appropriate, I think).