refactor, start real analyser attempts, make discoveries.
I know how to check a type matches an interface, now.  That's cool.

Is that the rabbithole I intended to go down first?
It absolutely is _not_.

However!  Turns out you can't easily pull in a dependency in go mod
to a package that's not published... so, my testcase files can't
actually import the error struct types in my root package...
So... ahheh, the "be able to recognize user-defined types" thing
went from far-down-the-road to a first step.  Okay!  Fun.
Let's goooooooo