Added fact for error types, recording error code and field information
This will have multiple benefits:
1. Faster processing due to memoization.
2. Being able to analyse errors that have their type declaration in
   other packages.
3. Other analysers can use this fact.

Still open: If another package defines an error type it currently has to be returned by an analysed function or the fact will not get created (and exported). We probably want to find error types automatically somehow or have them marked for analysis by a comment.