Finding error codes assigned in methods of error types
Now finding constant error codes that are assigned to the error code
field inside of methods of error types. Found codes are added to the
error type and no checks are implemented to test if the methods are
actually called.

This change was mainly implemented to find invalid assignments to the
error code field inside methods of the error type.