freshen up some documentation.
There's usable stuff now; let's talk about it as such!

I split the spec in half.  I'm not sure if that's an increase
to clarity or not.  The reason I split it is that the analysis
tool has evolved to have a strong opinion about the code thing,
and not about the rest -- so, to accurately describe what parts
are the minimum viable for adoption, we have to say something
about that.  I still think the rest of the suggested serialization
is worth touting, too, though.

There's more work to do here iteratively on docs, it seems.
Plenty of it.  This just takes one step forward.